requirements.txt
fastapi>=0.115.5
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.12.0
python-dotenv>=1.0.1
requests>=2.32.3
beautifulsoup4>=4.12.3
lxml>=5.3.0
lxml-html-clean>=0.4.1
newspaper3k==0.2.8
readability-lxml>=0.8.1
ffmpeg-python>=0.2.0
nltk>=3.9.1
google-generativeai>=0.8.3
pyttsx3>=2.98
gTTS>=2.5.3
aiofiles>=24.1.0
Pillow>=11.0.0
httpx>=0.27.2
